debug: move panic logic into kernel
This was previously just copied into each board, makes more sense
to centralize the heavy-lifting.
