+++ boards/imix/src/io.rs
@@ -1,5 +1,6 @@
 use core::fmt::*;
-use kernel::{debug, process};
+use kernel::debug;
+use kernel::hil::led;
 use kernel::hil::uart::{self, UART};
 use sam4l;
 
@@ -35,69 +36,9 @@ impl Write for Writer {
 #[no_mangle]
 #[lang = "panic_fmt"]
 pub unsafe extern "C" fn panic_fmt(args: Arguments, file: &'static str, line: u32) -> ! {
-    // XXX Replace with something like kernel::begin_panic()
-    // XXX Maybe place that call at panic_fmt, as it's called first
-    // XXX Better to cancel the transaction rather than hope we wait long enough
-    // Let any outstanding uart DMA's finish
-    asm!("nop");
-    asm!("nop");
-    for _ in 0..200000 {
-        asm!("nop");
-    }
-    asm!("nop");
-    asm!("nop");
-
+    let led = &mut led::LedLow::new(&mut sam4l::gpio::PC[10]);
     let writer = &mut WRITER;
-    let _ = writer.write_fmt(format_args!(
-        "\r\n\nKernel panic at {}:{}:\r\n\t\"",
-        file, line
-    ));
-    let _ = write(writer, args);
-    let _ = writer.write_str("\"\r\n");
-
-    // Print version of the kernel
-    let _ = writer.write_fmt(format_args!(
-        "\tKernel version {}\r\n",
-        env!("TOCK_KERNEL_VERSION")
-    ));
-
-    // Flush debug buffer if needed
-    debug::flush(writer);
-
-    // Print fault status once
-    let procs = &mut process::PROCS;
-    if procs.len() > 0 {
-        procs[0].as_mut().map(|process| {
-            process.fault_str(writer);
-        });
-    }
-
-    // print data about each process
-    let _ = writer.write_fmt(format_args!("\r\n---| App Status |---\r\n"));
-    let procs = &mut process::PROCS;
-    for idx in 0..procs.len() {
-        procs[idx].as_mut().map(|process| {
-            process.statistics_str(writer);
-        });
-    }
-
-    // blink the panic signal
-    let led = &sam4l::gpio::PC[10];
-    led.enable_output();
-    loop {
-        for _ in 0..1000000 {
-            led.clear();
-        }
-        for _ in 0..100000 {
-            led.set();
-        }
-        for _ in 0..1000000 {
-            led.clear();
-        }
-        for _ in 0..500000 {
-            led.set();
-        }
-    }
+    debug::panic(led, writer, args, file, line)
 }
 
 #[macro_export]
